+++ js/admin.js
@@ -1,3 +1,3 @@
 (function() {
-	'use strict';
+	// do nothing
 })();